// src/app/models/enums/form-field-names/generate-quotation-pdf-form-field.ts
export enum GenerateQuotationPdfFormField {
  PAYMENT_TERMS = "paymentTerms",
  PREPARED_BY = "preparedBy",
  APPROVED_BY = "approvedBy",
  GLOBAL_COMMENT = "globalComment",
  LINE_EXTRA_INFOS = "lineExtraInfos",
}

export enum QuotationLineExtraInfoFormField {
  LEAD_TIME_VALUE = "leadTimeValue",
  LEAD_TIME_UNIT = "leadTimeUnit",
  REMARKS = "remarks",
}